+++ src/planx/cellular_automata/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""PlanX Cellular Automata Submodule."""
+
+from .growth import markov_transition_probability_matrix, sleuth_cellular_automata_growth
+
+__all__ = [
+    "sleuth_cellular_automata_growth",
+    "markov_transition_probability_matrix",
+]