+++ src/planx/urban_physics/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""PlanX Urban Physics Submodule."""
+
+from .canopy import frontal_area_index_canopy, surface_albedo_cooling_potential
+
+__all__ = [
+    "frontal_area_index_canopy",
+    "surface_albedo_cooling_potential",
+]